+++ tests/lib.rs
@@ -10,8 +10,8 @@ mod errors;
 mod instrumentation;
 #[cfg(feature = "migrations")]
 mod migrations;
-mod notifications;
 mod multiconnection;
+mod notifications;
 #[cfg(any(feature = "bb8", feature = "deadpool", feature = "mobc"))]
 mod pooling;
 #[cfg(feature = "async-connection-wrapper")]